V 1.5.16: Reconfigure how default kinetic params are handled.
The values are not applied to the exchange reaction parameter vectors
(Vmax, KM, Hill Coeff, Alpha, W) so that a difference is maintained
between what is set to 'default' and what has a specific value. A
negative value in the vector means to use the model's default. A
negative value in the model's default means to use the Package default.

Also added a checkbox to the UI to select whether to override the
models' defaults.