fix bug in FBAOptimizerGurobi
The previous implementation crashed if the model contained a metabolite
not connected to any reactions. While this definetaly is weird, it may
happen and shouldn't cause COMETS to crash.

I've commented out a few lines with the old code and added an extra loop
to solve this in an efficient and more stable way.