Improve light modelling
I have realized that light needs to be accounted for differently when
calculating maximal uptake rates for the model. Theere are two main
differences:
1) The light is not a concentration similar to other media components,
but rather a flux [mmol photons / m^2/ s].
2) The uptake of photons is determined by the surface area of the
organism and the absorption coefficient

The maximum uptake rate is determined by the surface area of the grid
cell and the photon flux. Also, because photons doesn't really hang
around it doesn't really make sense to update the media. However, the
user should be careful if multiple species are modelled, because
"competition" of the light as a resource is not handled.