Bug fix - stationnary phase was turning on when it didn't have to
In order to save computational time in simulations with many models and batchDilution,
a "stationnary phase" flag was implemented. When no models are growing, this flag
turns on which prevents any model from optimizing, as long as the environment is not
externally changed (through a batch transfer). However, sometimes obligate crossfeeders
can be using media and secreting to the environment without growing at all (just keeping
ATP maintenance).

In the new version, the algorithm now checks for changes in the environment instead of
the models' biomass, and turns on stationnary phase only if the environment does not
change at all in a given cycle.